requirements.txt
streamlit==1.42.2
google-generativeai==0.8.4
openai-whisper==20240930
python-dotenv==1.0.1
torch==2.0.0
torchaudio==2.0.0
numpy==1.23.5
soundfile==0.13.1
openai==1.63.2
streamlit-mic-recorder==0.0.8
Pillow==11.1.0
